+++ src/Infrastructure/Dto/Order/OrderResponse.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace App\Infrastructure\Dto\Order;
 
 use App\Domain\Order\Entity\Order;
-use App\Domain\Thing\Dto\ThingOut;
+use App\Infrastructure\Dto\Thing\ThingResponse;
 use App\Domain\User\Dto\RequesterOut;
 use Nelmio\ApiDocBundle\Annotation\Model;
 use Swagger\Annotations as SWG;
@@ -16,8 +16,8 @@ class OrderResponse
     public string $id;
     /** @SWG\Property(type=@SWG\Schema(@Model(type=RequesterOut::class))) */
     public RequesterOut $requester;
-    /** @SWG\Property(type=@SWG\Schema(@Model(type=ThingOut::class))) */
-    public ThingOut $thing;
+    /** @SWG\Property(type=@SWG\Schema(@Model(type=ThingResponse::class))) */
+    public ThingResponse $thing;
     /** @SWG\Property(type="integer") */
     public int $quantity;
     /** @SWG\Property(type="integer") */
@@ -29,7 +29,7 @@ class OrderResponse
 
         $self->id = $order->getId();
         $self->requester = RequesterOut::createFromRequester($order->getRequester());
-        $self->thing = ThingOut::createFromThing($order->getThing());
+        $self->thing = ThingResponse::createFromThing($order->getThing());
 
         $self->quantity = $order->getQuantity();
         $self->remaining = $order->getRemaining();
